+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Bump rewrite-clj to `1.1.48`.
   - Bump cljfmt to `1.13.0`.
   - Bump babashka/fs to `0.5.22`.
+  - New feature: completion of lib versions on deps.edn, bb.edn and project.clj files. #1905
 
 ## 2024.08.05-18.16.00